+++ src/command_modules/azure-cli-sql/azure/cli/command_modules/sql/_params.py
@@ -179,10 +179,17 @@ def load_arguments(self, _):
         c.argument('usage_name', options_list=['--usage', '-u'])
 
     with self.argument_context('sql db') as c:
+        c.argument('server_name',
+                   arg_type=server_param_type,
+                   # Allow --ids command line argument. id_part=name is 1st name in uri
+                   id_part='name')
+
         c.argument('database_name',
                    options_list=['--name', '-n'],
-                   help='Name of the Azure SQL Database.')
-        c.argument('server_name', arg_type=server_param_type)
+                   help='Name of the Azure SQL Database.',
+                   # Allow --ids command line argument. id_part=child_name_1 is 2nd name in uri
+                   id_part='child_name_1')
+
         c.argument('elastic_pool_name', options_list=['--elastic-pool'])
         c.argument('requested_service_objective_name', options_list=['--service-objective'])
         c.argument('max_size_bytes', options_list=['--max-size'],
@@ -500,10 +507,15 @@ def load_arguments(self, _):
     #                sql dw                       #
     ###############################################
     with self.argument_context('sql dw') as c:
-        c.argument('database_name', options_list=['--name', '-n'],
-                   help='Name of the data warehouse.')
+        c.argument('server_name',
+                   arg_type=server_param_type,
+                   # Allow --ids command line argument. id_part=name is 1st name in uri
+                   id_part='name')
 
-        c.argument('server_name', arg_type=server_param_type)
+        c.argument('database_name', options_list=['--name', '-n'],
+                   help='Name of the data warehouse.',
+                   # Allow --ids command line argument. id_part=child_name_1 is 2nd name in uri
+                   id_part='child_name_1')
 
         c.argument('max_size_bytes', options_list=['--max-size'],
                    type=SizeWithUnitConverter('B', result_type=int),
@@ -539,10 +551,16 @@ def load_arguments(self, _):
         # This line to be removed when zone redundancy is fully supported in production.
         c.ignore('zone_redundant')
 
+        c.argument('server_name',
+                   arg_type=server_param_type,
+                   # Allow --ids command line argument. id_part=name is 1st name in uri
+                   id_part='name')
+
         c.argument('elastic_pool_name',
                    options_list=['--name', '-n'],
-                   help='The name of the elastic pool.')
-        c.argument('server_name', arg_type=server_param_type)
+                   help='The name of the elastic pool.',
+                   # Allow --ids command line argument. id_part=child_name_1 is 2nd name in uri
+                   id_part='child_name_1')
 
         # --db-dtu-max and --db-dtu-min were the original param names, which is consistent with the
         # underlying REST API.
@@ -598,7 +616,9 @@ def load_arguments(self, _):
     #                sql server                   #
     ###############################################
     with self.argument_context('sql server') as c:
-        c.argument('server_name', options_list=['--name', '-n'])
+        c.argument('server_name', options_list=['--name', '-n'],
+                   # Allow --ids command line argument. id_part=name is 1st name in uri
+                   id_part='name')
         c.argument('administrator_login', options_list=['--admin-user', '-u'])
         c.argument('administrator_login_password', options_list=['--admin-password', '-p'])
 
@@ -659,11 +679,16 @@ def load_arguments(self, _):
     with self.argument_context('sql server firewall-rule') as c:
         # Help text needs to be specified because 'sql server firewall-rule update' is a custom
         # command.
-        c.argument('server_name', arg_type=server_param_type)
+        c.argument('server_name',
+                   arg_type=server_param_type,
+                   # Allow --ids command line argument. id_part=name is 1st name in uri
+                   id_part='name')
 
         c.argument('firewall_rule_name',
                    options_list=['--name', '-n'],
-                   help='The name of the firewall rule.')
+                   help='The name of the firewall rule.',
+                   # Allow --ids command line argument. id_part=child_name_1 is 2nd name in uri
+                   id_part='child_name_1')
 
         c.argument('start_ip_address',
                    options_list=['--start-ip-address'],